databricks-functions-dataexplorer/functions/Function-DataExplorer-Ingestion/__app__/requirements.txt
applicationinsights==0.11.9
azure-common==1.1.25
azure-core==1.6.0
azure-cosmosdb-table==1.0.6
azure-storage-blob==12.5.0
azure-storage-queue==12.1.3
azure-functions==1.2.1
azure-kusto-data==1.0.2
azure-kusto-ingest==1.0.2
python-dateutil==2.8.1
tenacity==6.2.0